Layouts is now a drop down link and moved ApplicationLayout under it
